# Remove non-layout style and prop updates path via `synchronouslyUpdatePropsOnUIThread`