--- a/packages/react-native-reanimated/android/src/paper/java/com/swmansion/reanimated/ReaCompatibility.java
+++ b/packages/react-native-reanimated/android/src/paper/java/com/swmansion/reanimated/ReaCompatibility.java
@@ -1,7 +1,6 @@
 package com.swmansion.reanimated;
 
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReadableMap;
 
 class ReaCompatibility {
   public ReaCompatibility(ReactApplicationContext reactApplicationContext) {}