--- a/packages/react-native-reanimated/apple/reanimated/apple/native/PlatformDepMethodsHolderImpl.mm
+++ b/packages/react-native-reanimated/apple/reanimated/apple/native/PlatformDepMethodsHolderImpl.mm
@@ -102,18 +102,6 @@ RequestRenderFunction makeRequestRender(REANodesManager *nodesManager)
 }
 
 #ifdef RCT_NEW_ARCH_ENABLED
-SynchronouslyUpdateUIPropsFunction makeSynchronouslyUpdateUIPropsFunction(REANodesManager *nodesManager)
-{
-  auto synchronouslyUpdateUIPropsFunction = [nodesManager](jsi::Runtime &rt, Tag tag, const jsi::Object &props) {
-    NSNumber *viewTag = @(tag);
-    NSDictionary *uiProps = convertJSIObjectToNSDictionary(rt, props);
-    [nodesManager synchronouslyUpdateViewOnUIThread:viewTag props:uiProps];
-  };
-  return synchronouslyUpdateUIPropsFunction;
-}
-#endif // RCT_NEW_ARCH_ENABLED
-
-#ifdef RCT_NEW_ARCH_ENABLED
 // nothing
 #else // RCT_NEW_ARCH_ENABLED
 UpdatePropsFunction makeUpdatePropsFunction(REAModule *reaModule)
@@ -291,10 +279,6 @@ makePlatformDepMethodsHolder(RCTBridge *bridge, REANodesManager *nodesManager, R
   auto requestRender = makeRequestRender(nodesManager);
 
 #ifdef RCT_NEW_ARCH_ENABLED
-  auto synchronouslyUpdateUIPropsFunction = makeSynchronouslyUpdateUIPropsFunction(nodesManager);
-#endif // RCT_NEW_ARCH_ENABLED
-
-#ifdef RCT_NEW_ARCH_ENABLED
   // nothing
 #else
   RCTUIManager *uiManager = nodesManager.uiManager;
@@ -354,7 +338,7 @@ makePlatformDepMethodsHolder(RCTBridge *bridge, REANodesManager *nodesManager, R
   PlatformDepMethodsHolder platformDepMethodsHolder = {
       requestRender,
 #ifdef RCT_NEW_ARCH_ENABLED
-      synchronouslyUpdateUIPropsFunction,
+      // nothing
 #else
       updatePropsFunction,
       scrollToFunction,
@@ -384,8 +368,6 @@ PlatformDepMethodsHolder makePlatformDepMethodsHolderBridgeless(
 {
   auto requestRender = makeRequestRender(nodesManager);
 
-  auto synchronouslyUpdateUIPropsFunction = makeSynchronouslyUpdateUIPropsFunction(nodesManager);
-
   auto getAnimationTimestamp = makeGetAnimationTimestamp();
 
   auto progressLayoutAnimation = makeProgressLayoutAnimation(reaModule);
@@ -410,7 +392,6 @@ PlatformDepMethodsHolder makePlatformDepMethodsHolderBridgeless(
 
   PlatformDepMethodsHolder platformDepMethodsHolder = {
       requestRender,
-      synchronouslyUpdateUIPropsFunction,
       getAnimationTimestamp,
       progressLayoutAnimation,
       endLayoutAnimation,