@tomekzaw 2025-02-12
> should we also merge/rework props lists in TS?

We'll need to discuss this. There's no need to keep separate sets for UI and native props (since I've merged the logic of applying UI and native props) but I think we still need to expose `addWhitelistedNativeProps` and `addWhitelistedUIProps` for backward-compatibility of third-party libraries built on top on Reanimated, to some extent.