@tjzel 2025-02-14
@tomekzaw We can keep both these methods and mark one of them as deprecated.